Fix an issue with job configurations where job configs would 'inherit' the configuration of other job files, if they didn't already have one and a new job being configured.  Fix an issue with job configurations when new jobs that were created (sample jobs) wouldn't have their sample file auto-selected for the user.  Fix an issue with job configurations where new jobs that were created would inherit previously viewed job data.  Ideally, now that we store interim job data on the server, we ought not to maintain the job config state in the browser, but rather push it down to the server to be stored.  This isn't a small change, so perhaps Jeremy ought to 'weigh in' on it before/if I try to implement it.
Also implement a fix where we cascade deletes of jobs down to their corresponding job files, assuming those files aren't being used for other jobs.  The current implementation retrieves a list of output files, then matches them with input files (for other jobs) and shows the checkbox if one or more of the output files are not being used for input files, then it deletes the job and files - if the checkbox (which defaults to checked) is checked.  There's a small UI display issue that is proving difficult to fix (in progress), but all the other (code) work is complete on this.